Fix #46 - Split showdown extensions
