+++ src/backend/extensions/index.ts
@@ -1,0 +1,2 @@
+export * from './headingAnchorExt'
+export * from './codeBlockExt'
